Change ecma_length_t and jerry_api_length_t from uint16_t to uint32_t.
JerryScript-DCO-1.0-Signed-off-by: Andrey Shitov a.shitov@samsung.com

+++ jerry-core/ecma/builtin-objects/ecma-builtins.cpp
@@ -601,7 +601,7 @@ ecma_builtin_bin_search_for_magic_string_id_in_array (const lit_magic_string_id_
 #endif /* !JERRY_NDEBUG */
 
   int32_t min = 0;
-  int32_t max = array_length - 1;
+  int32_t max = (int32_t) array_length - 1;
 
   while (min <= max)
   {
